refactor(parser): allow parser to toggle between scanners

+++ libflux/core/src/scanner/mod.rs
@@ -154,12 +154,12 @@ impl Scanner {
         self.checkpoint_line = self.cur_line;
         self.checkpoint_last_newline = self.last_newline;
 
-        let mut token_start = 0 as u32;
-        let mut token_start_line = 0 as u32;
-        let mut token_start_col = 0 as u32;
-        let mut token_end = 0 as u32;
-        let mut token_end_line = 0 as u32;
-        let mut token_end_col = 0 as u32;
+        let mut token_start = 0_u32;
+        let mut token_start_line = 0_u32;
+        let mut token_start_col = 0_u32;
+        let mut token_end = 0_u32;
+        let mut token_end_line = 0_u32;
+        let mut token_end_col = 0_u32;
 
         let error = unsafe {
             scan(
